+++ register-login/src/main/java/com/info/register_login/service/IUserService.java
@@ -16,7 +16,7 @@ public interface IUserService {
 	 * 根据id删除用户
 	 * @param user_id
 	 */
-	void deleteUserById(Integer user_id);
+	void deleteUserById(Long user_id);
 
 
 	/**
@@ -37,7 +37,7 @@ public interface IUserService {
 	 * @param user_id
 	 * @return
 	 */
-	public User findById(Integer user_id);
+	public User findById(Long user_id);
 
 
 	/**
